requirements.txt
Flask==3.0.0
boto3==1.34.0
serverless-wsgi==3.0.3
google-auth==2.25.2
requests==2.31.0
